+++ src/Application/Query/Sheet/Detail/ParticipantDetailQuery.php
@@ -10,6 +10,7 @@
 
 namespace Proximum\Vimeet\Application\Query\Sheet\Detail;
 
+use Proximum\Vimeet\Domain\Model\Admin;
 use Proximum\Vimeet\Domain\Model\Event;
 use Proximum\Vimeet\Domain\Model\Sheet;
 
@@ -24,16 +25,14 @@ class ParticipantDetailQuery
     /** @var string */
     public $locale;
 
-    /**
-     * ParticipantDetailQuery constructor.
-     *
-     * @param Sheet  $sheet
-     * @param string $locale
-     */
-    public function __construct(Sheet $sheet, string $locale)
+    /** @var Admin */
+    public $admin;
+
+    public function __construct(Admin $admin, Sheet $sheet, string $locale)
     {
         $this->event  = $sheet->getEvent();
         $this->sheet  = $sheet;
         $this->locale = $locale;
+        $this->admin = $admin;
     }
 }