+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - MV-182 - Show amount included taxes when event does not have taxes
 ### Added
 - VIMEET-1889 - Take account of type rules when there is no categories rules
+- VIMEET-2019 - View multiple sheets contacts list; Impersonate as every user of a participation sheet
 
 ## [2.11.0] - 2019-08-27
 ### Added